SIH-PM-INTERNSHIP/backend/models/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
requests==2.32.3
pdfminer.six==20240706
pytesseract==0.3.13
Pillow==10.4.0
spacy==3.7.6

# Optional: install English model separately
# python -m spacy download en_core_web_sm

